[themotte/rDrama#451] Move the pusher_thread() function from files/routes/comment.py to files/helpers/comment.py
